Move format enum into its own module
